+++ src/components/pages/Home/Home.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import { CiUser, CiRainbow, CiTwitter, CiSun } from "react-icons/ci";
-import { useScrollToTop } from "hooks/scroll";
 import BUTTON_COLORS from "constants/ButtonColors";
 import BUTTON_VARIANTS from "constants/ButtonVariants";
 
@@ -45,126 +44,120 @@ function buildProducts() {
   }));
 }
 
-const Home = () => {
-  useScrollToTop();
-  return (
-    <>
-      <Hero image={BgBrushImage}>
-        <Heading>
-          <h1>
-            Inspire um <strong>mundo</strong> mais
-            <br />
-            colorido
-          </h1>
-        </Heading>
-        <ul>
-          <li>Lorem ipsum dolor sit amet, consectetur adipiscing.</li>
-          <li>Lorem ipsum dolor sit amet.</li>
-          <li>Lorem ipsum dolor sit amet, consectetur.</li>
-        </ul>
-        <Button
-          color={BUTTON_COLORS.PRIMARY}
-          variant={BUTTON_VARIANTS.OUTLINED}
-        >
-          Matricule-se agora
-        </Button>
-      </Hero>
-      <Section>
-        <Grid sm={2} md={4}>
-          <Feature icon={<CiUser />} title="Pessoas e detalhes da face">
-            <p>
-              What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
-              printing and typesetting industry. Lorem Ipsum has been the
-              industry's standard dummy text ever since the 1500s.
-            </p>
-          </Feature>
-          <Feature icon={<CiRainbow />} title="Paisagens e horizontes">
-            <p>
-              What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
-              printing and typesetting industry. Lorem Ipsum has been the
-              industry's standard dummy text ever since the 1500s.
-            </p>
-          </Feature>
-          <Feature icon={<CiTwitter />} title="Animais">
-            <p>
-              What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
-              printing and typesetting industry. Lorem Ipsum has been the
-              industry's standard dummy text ever since the 1500s.
-            </p>
-          </Feature>
-          <Feature icon={<CiSun />} title="Céu">
-            <p>
-              What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
-              printing and typesetting industry. Lorem Ipsum has been the
-              industry's standard dummy text ever since the 1500s.
-            </p>
-          </Feature>
-        </Grid>
-      </Section>
-      <Section inverse>
-        <Heading>
-          <h2>Conteúdo das aulas</h2>
-        </Heading>
-        <ProductGrid products={buildProducts()} />
-      </Section>
-      <Section>
-        <Grid md={2}>
-          <div>
-            <Heading>
-              <h2>Quem sou eu</h2>
-            </Heading>
-            <p>
-              What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
-              printing and typesetting industry. Lorem Ipsum has been the
-              industry's standard dummy text ever since the 1500s.
-            </p>
-            <Button as={Link} to="/sobre" color={BUTTON_COLORS.PRIMARY}>
-              Saiba mais
-            </Button>
-          </div>
-          <div>
-            <video
-              src={WatercolorVideo}
-              width="100%"
-              autoPlay
-              playsInline
-              loop
-              muted
-            />
-          </div>
-        </Grid>
-      </Section>
-      <Section inverse>
-        <Heading>
-          <h2>Dúvidas frequentes</h2>
-        </Heading>
-        <AccordionGroup>
-          <Accordion title="Quando as aula começam?">
-            <p>
-              What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
-              printing and typesetting industry. Lorem Ipsum has been the
-              industry's
-            </p>
-          </Accordion>
-          <Accordion title="Não tenho todos os materiais, o que faço?">
-            <p>
-              What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
-              printing and typesetting industry. Lorem Ipsum has been the
-              industry's
-            </p>
-          </Accordion>
-          <Accordion title="Como posso realizar o pagamento?">
-            <p>
-              What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
-              printing and typesetting industry. Lorem Ipsum has been the
-              industry's
-            </p>
-          </Accordion>
-        </AccordionGroup>
-      </Section>
-      <Footer />
-    </>
-  );
-};
+const Home = () => (
+  <>
+    <Hero image={BgBrushImage}>
+      <Heading>
+        <h1>
+          Inspire um <strong>mundo</strong> mais
+          <br />
+          colorido
+        </h1>
+      </Heading>
+      <ul>
+        <li>Lorem ipsum dolor sit amet, consectetur adipiscing.</li>
+        <li>Lorem ipsum dolor sit amet.</li>
+        <li>Lorem ipsum dolor sit amet, consectetur.</li>
+      </ul>
+      <Button color={BUTTON_COLORS.PRIMARY} variant={BUTTON_VARIANTS.OUTLINED}>
+        Matricule-se agora
+      </Button>
+    </Hero>
+    <Section>
+      <Grid sm={2} md={4}>
+        <Feature icon={<CiUser />} title="Pessoas e detalhes da face">
+          <p>
+            What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
+            printing and typesetting industry. Lorem Ipsum has been the
+            industry's standard dummy text ever since the 1500s.
+          </p>
+        </Feature>
+        <Feature icon={<CiRainbow />} title="Paisagens e horizontes">
+          <p>
+            What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
+            printing and typesetting industry. Lorem Ipsum has been the
+            industry's standard dummy text ever since the 1500s.
+          </p>
+        </Feature>
+        <Feature icon={<CiTwitter />} title="Animais">
+          <p>
+            What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
+            printing and typesetting industry. Lorem Ipsum has been the
+            industry's standard dummy text ever since the 1500s.
+          </p>
+        </Feature>
+        <Feature icon={<CiSun />} title="Céu">
+          <p>
+            What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
+            printing and typesetting industry. Lorem Ipsum has been the
+            industry's standard dummy text ever since the 1500s.
+          </p>
+        </Feature>
+      </Grid>
+    </Section>
+    <Section inverse>
+      <Heading>
+        <h2>Conteúdo das aulas</h2>
+      </Heading>
+      <ProductGrid products={buildProducts()} />
+    </Section>
+    <Section>
+      <Grid md={2}>
+        <div>
+          <Heading>
+            <h2>Quem sou eu</h2>
+          </Heading>
+          <p>
+            What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
+            printing and typesetting industry. Lorem Ipsum has been the
+            industry's standard dummy text ever since the 1500s.
+          </p>
+          <Button as={Link} to="/sobre" color={BUTTON_COLORS.PRIMARY}>
+            Saiba mais
+          </Button>
+        </div>
+        <div>
+          <video
+            src={WatercolorVideo}
+            width="100%"
+            autoPlay
+            playsInline
+            loop
+            muted
+          />
+        </div>
+      </Grid>
+    </Section>
+    <Section inverse>
+      <Heading>
+        <h2>Dúvidas frequentes</h2>
+      </Heading>
+      <AccordionGroup>
+        <Accordion title="Quando as aula começam?">
+          <p>
+            What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
+            printing and typesetting industry. Lorem Ipsum has been the
+            industry's
+          </p>
+        </Accordion>
+        <Accordion title="Não tenho todos os materiais, o que faço?">
+          <p>
+            What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
+            printing and typesetting industry. Lorem Ipsum has been the
+            industry's
+          </p>
+        </Accordion>
+        <Accordion title="Como posso realizar o pagamento?">
+          <p>
+            What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the
+            printing and typesetting industry. Lorem Ipsum has been the
+            industry's
+          </p>
+        </Accordion>
+      </AccordionGroup>
+    </Section>
+    <Footer />
+  </>
+);
 
 export default Home;
