+++ src/components/pages/Home/Home.test.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { getByRole, render } from "test-utils";
+import { render } from "test-utils";
 import Home from "./Home";
 
 describe("Home Page", () => {
@@ -21,5 +21,10 @@ describe("Home Page", () => {
     expect(getByText("Conteúdo das aulas")).toBeInTheDocument();
     expect(getByText("Quem sou eu")).toBeInTheDocument();
     expect(getByText("Dúvidas frequentes")).toBeInTheDocument();
+  });
+
+  it("should scrolls to top in first render", () => {
+    mount();
+    expect(window.scrollTo).toBeCalledTimes(1);
   });
 });
