+++ src/components/pages/Product/ProductDetails.js
@@ -3,8 +3,6 @@ import PropTypes from "prop-types";
 import styled from "styled-components";
 import { FaIdCard, FaScroll, FaHome } from "react-icons/fa";
 
-import { useScrollToTop } from "hooks/scroll";
-
 import Callout from "components/atoms/Callout";
 import Heading from "components/atoms/Heading";
 import Hero from "components/molecules/Hero";
@@ -39,62 +37,58 @@ const items = [
   { label: "Nome do serviço" },
 ];
 
-const ProductDetails = () => {
-  useScrollToTop();
-
-  return (
-    <>
-      <Hero image={BgAboutImage}>
-        <Heading>
-          <h1> Nome do serviço</h1>
-        </Heading>
-        <BreadCrumb items={items} />
-      </Hero>
-      <Section>
-        <p>
-          Lorem Ipsum is simply dummy text of the printing and typesetting
-          industry.
-        </p>
-        <p>
-          Contrary to popular belief, Lorem Ipsum is not simply random text. It
-          has roots in a piece of classical.
-        </p>
-        <p>
-          There are many variations of passages of Lorem Ipsum available, but
-          the majority.
-        </p>
-        <h5>Documentos necessários:</h5>
-        <PinnedList>
-          <PinnedItem>
-            <FaIdCard />
-            RG
-          </PinnedItem>
-          <PinnedItem>
-            <FaIdCard />
-            CPF
-          </PinnedItem>
-          <PinnedItem>
-            <FaScroll />
-            Certidão de nascimento
-          </PinnedItem>
-          <PinnedItem>
-            <FaHome />
-            Comprovante de residência
-          </PinnedItem>
-        </PinnedList>
-      </Section>
-      <Section inverse>
-        <Callout
-          title={"Quer ganhar um desconto nesse curso?"}
-          description={`What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s.`}
-          svg={<Art />}
-          buttonDescription="Saiba mais"
-        />
-      </Section>
-      <Footer />
-    </>
-  );
-};
+const ProductDetails = () => (
+  <>
+    <Hero image={BgAboutImage}>
+      <Heading>
+        <h1> Nome do serviço</h1>
+      </Heading>
+      <BreadCrumb items={items} />
+    </Hero>
+    <Section>
+      <p>
+        Lorem Ipsum is simply dummy text of the printing and typesetting
+        industry.
+      </p>
+      <p>
+        Contrary to popular belief, Lorem Ipsum is not simply random text. It
+        has roots in a piece of classical.
+      </p>
+      <p>
+        There are many variations of passages of Lorem Ipsum available, but the
+        majority.
+      </p>
+      <h5>Documentos necessários:</h5>
+      <PinnedList>
+        <PinnedItem>
+          <FaIdCard />
+          RG
+        </PinnedItem>
+        <PinnedItem>
+          <FaIdCard />
+          CPF
+        </PinnedItem>
+        <PinnedItem>
+          <FaScroll />
+          Certidão de nascimento
+        </PinnedItem>
+        <PinnedItem>
+          <FaHome />
+          Comprovante de residência
+        </PinnedItem>
+      </PinnedList>
+    </Section>
+    <Section inverse>
+      <Callout
+        title={"Quer ganhar um desconto nesse curso?"}
+        description={`What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s.`}
+        svg={<Art />}
+        buttonDescription="Saiba mais"
+      />
+    </Section>
+    <Footer />
+  </>
+);
 
 ProductDetails.propTypes = {};
