+++ src/routes/home.js
@@ -1,0 +1,10 @@
+import React from "react";
+import { useScrollToTop } from "hooks/scroll";
+import HomePage from "components/pages/Home";
+
+const Home = () => {
+  useScrollToTop();
+  return <HomePage />;
+};
+
+export default Home;
