+++ src/routes/index.js
@@ -1,0 +1,16 @@
+import React from "react";
+import { Routes, Route } from "react-router-dom";
+
+import Home from "routes/home";
+import About from "routes/about";
+import ProductDetails from "routes/products";
+
+const AppRoutes = () => (
+  <Routes>
+    <Route path="/" element={<Home />} />
+    <Route path="/sobre" element={<About />} />
+    <Route path="/servicos" element={<ProductDetails />} />
+  </Routes>
+);
+
+export default AppRoutes;
